--- libs/common/project.json
@@ -1,33 +1,0 @@
-{
-  "name": "@foblex/mr-common",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/common/src",
-  "prefix": "lib",
-  "projectType": "library",
-  "release": {
-    "version": {
-      "manifestRootsToUpdate": ["dist/{projectRoot}"],
-      "currentVersionResolver": "git-tag",
-      "fallbackCurrentVersionResolver": "disk"
-    }
-  },
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nx/angular:package",
-      "outputs": ["{workspaceRoot}/dist/{projectRoot}"],
-      "options": {
-        "project": "libs/common/ng-package.json"
-      },
-      "configurations": {
-        "production": {
-          "tsConfig": "libs/common/tsconfig.lib.prod.json"
-        },
-        "development": {
-          "tsConfig": "libs/common/tsconfig.lib.json"
-        }
-      },
-      "defaultConfiguration": "production"
-    }
-  }
-}
