--- libs/common/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/utils';
-
