--- libs/highlight/eslint.config.mjs
@@ -1,5 +1,0 @@
-import baseConfig from '../../eslint.config.mjs';
-
-export default [
-  ...baseConfig,
-];
