--- libs/highlight/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/libs/mr-highlight",
-  "lib": {
-    "entryFile": "src/index.ts"
-  }
-}
