--- libs/highlight/project.json
@@ -1,33 +1,0 @@
-{
-  "name": "@foblex/mr-highlight",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/highlight/src",
-  "prefix": "lib",
-  "projectType": "library",
-  "release": {
-    "version": {
-      "manifestRootsToUpdate": ["dist/{projectRoot}"],
-      "currentVersionResolver": "git-tag",
-      "fallbackCurrentVersionResolver": "disk"
-    }
-  },
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nx/angular:package",
-      "outputs": ["{workspaceRoot}/dist/{projectRoot}"],
-      "options": {
-        "project": "libs/highlight/ng-package.json"
-      },
-      "configurations": {
-        "production": {
-          "tsConfig": "libs/highlight/tsconfig.lib.prod.json"
-        },
-        "development": {
-          "tsConfig": "libs/highlight/tsconfig.lib.json"
-        }
-      },
-      "defaultConfiguration": "production"
-    }
-  }
-}
