+++ libs/m-render/src/index.ts
@@ -1,6 +1,7 @@
 export * from './lib/analytics';
 export * from './lib/common';
 export * from './lib/documentation-page';
+export * from './lib/highlight';
 export * from './lib/home-page';
 export * from './lib/theme';
