+++ libs/m-render/src/lib/common/index.ts
@@ -2,3 +2,4 @@ export * from './components';
 export * from './directives';
 export * from './domain';
 export * from './services';
+export * from './utils';
