+++ libs/m-render/src/lib/common/utils/window.ts
