+++ libs/m-render/src/lib/documentation-page/components/markdown-container/f-markdown/components/f-code-group/f-code-view/components/highlight.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, DestroyRef, ElementRef, inject, input, OnInit } from '@angular/core';
 import { take } from 'rxjs/operators';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { HighlightService } from '@foblex/mr-highlight';
+import { HighlightService } from '../../../../../../../../highlight';
 
 @Component({
   selector: 'highlight',
