+++ libs/m-render/src/lib/documentation-page/components/markdown-container/f-markdown/components/f-code-group/f-code-view/f-code-view.component.ts
@@ -6,9 +6,8 @@ import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { HighlightComponent } from './components/highlight.component';
 import { parseLanguageFromFileExtension, parseSyntaxLanguage } from './utils/parse-language';
 import { IParsedContainerData } from '../../../markdown';
-import { PopoverService } from '../../../../../../../common';
+import { copyToClipboard, PopoverService } from '../../../../../../../common';
 import { coerceComponentHeight } from '../f-example-view';
-import { copyToClipboard } from '@foblex/mr-common';
 
 @Component({
   selector: 'f-code-view',
