+++ libs/m-render/src/lib/documentation-page/components/markdown-container/f-markdown/components/f-markdown-footer/f-markdown-footer.component.ts
@@ -15,7 +15,7 @@ import {
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { DocumentationStore } from '../../../../../services';
 import { HandleNavigationLinksHandler, HandleNavigationLinksRequest } from '../../../../../domain';
-import { WINDOW } from '@foblex/mr-common';
+import { WINDOW } from '../../../../../../common';
 
 @Component({
   selector: 'footer [f-markdown-footer]',
