+++ libs/m-render/src/lib/documentation-page/components/markdown-container/f-markdown/f-markdown-renderer.component.ts
@@ -27,7 +27,7 @@ import {
 } from '../../scrollable-container';
 import { FMarkdownFooterComponent } from './components';
 import { HandleNavigationLinksHandler, HandleNavigationLinksRequest } from '../../../domain';
-import { IS_BROWSER_PLATFORM, WINDOW } from '@foblex/mr-common';
+import { IS_BROWSER_PLATFORM, WINDOW } from '../../../../common';
 
 @Component({
   selector: 'f-markdown-renderer',
