+++ libs/m-render/src/lib/documentation-page/components/scrollable-container/domain/calculate-hash-from-scroll-position-and-activate-toc/calculate-hash-from-scroll-position-and-activate-toc.handler.ts
@@ -6,7 +6,7 @@ import {
 } from '../../../index';
 import { ActivateTocByHashRequest } from '../activate-toc-by-hash';
 import { DocumentationStore } from '../../../../services';
-import { DOCUMENT_ELEMENT, WINDOW } from '@foblex/mr-common';
+import { DOCUMENT_ELEMENT, WINDOW } from '../../../../../common';
 
 interface IHasTopItem {
   hash: string;
